requirements.txt

# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

numpy
scikit-learn
joblib
scipy
setuptools
pyarrow
pyyaml
pandas
jupyter
xarray
netCDF4
pyproj
rasterio
geojson
rioxarray
sklearn
tensorflow-gpu
plotly
sparse
pmdarima
seaborn
hyperopt
binpacking
pydantic==1.10.1
#pandarallel
#swifter
#neuralhydrology
#GDAL
